requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2018.10.15
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
Django==3.0.8
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-filter==2.3.0
django-qr-code==1.2.0
django-qrcode==0.3
gunicorn==20.0.4
idna==2.7
importlib-metadata==1.7.0
importlib-resources==3.0.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
numpy==1.19.1
pandas==1.1.0
Pillow==5.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2018.5
qrcode==6.1
requests==2.19.1
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
sweetviz==1.0a7
tqdm==4.46.1
urllib3==1.23
whitenoise==5.1.0
wincertstore==0.2
zipp==3.1.0
